add support for externally created api-gw authorizers
Setting a valid `authorizationType` and `authorizerId` to the id of an
authorizer resource already belonging to the api-gateway will now configure the
authorizer to protect that endpoint

(abbreviated `s-function.json`)
```
{
  "name": "Ping",
  "runtime": "nodejs10.x",
  "endpoints": [
    {
      "authorizationType": "COGNITO_USER_POOLS",
      "authorizerId": "${authorizerID}",
    }
  ]
}

```

Note: this breaks the previous "magic" serverless behaviour that would create a
custom authorizer lambda function by specifiying `authorizer` key on the root
of the `s-function.json`. (but you probably weren't using that anyway).

Note II: This does not manage the lifecycle of the authorizer for you.
Note III: a method's authorizer applies to ALL stages and can't be configured
	differently for each stage on a given api-gw. To get around this use a
	seperate api-gw for each different authorizer.